// Source/Client/Interop/JSInteropTypes.Generated.ts
export interface IGraphicsInterop {
    initialize(containerId: string): Promise<void>;
    dispose(): Promise<void>;
    clear(): Promise<void>;
    getBackgroundColor(): Promise<string>;
    setBackgroundColor(backgroundColor: string): Promise<void>;
    getBrushColor(): Promise<string>;
    setBrushColor(brushColor: string): Promise<void>;
    getPenColor(): Promise<string>;
    setPenColor(penColor: string): Promise<void>;
    getPenWidth(): Promise<number>;
    setPenWidth(penWidth: number): Promise<void>;
    getFontBold(): Promise<boolean>;
    setFontBold(fontBold: boolean): Promise<void>;
    getFontItalic(): Promise<boolean>;
    setFontItalic(fontItalic: boolean): Promise<void>;
    getFontName(): Promise<string>;
    setFontName(fontName: string): Promise<void>;
    getFontSize(): Promise<number>;
    setFontSize(fontSize: number): Promise<void>;
    drawBoundText(x: number, y: number, width: number, text: string): Promise<void>;
    drawEllipse(x: number, y: number, width: number, height: number): Promise<void>;
    drawImage(imageName: string, x: number, y: number): Promise<void>;
    drawLine(x1: number, y1: number, x2: number, y2: number): Promise<void>;
    drawRectangle(x: number, y: number, width: number, height: number): Promise<void>;
    drawResizedImage(imageName: string, x: number, y: number, width: number, height: number): Promise<void>;
    drawText(x: number, y: number, text: string): Promise<void>;
    drawTriangle(x1: number, y1: number, x2: number, y2: number, x3: number, y3: number): Promise<void>;
    fillEllipse(x: number, y: number, width: number, height: number): Promise<void>;
    fillRectangle(x: number, y: number, width: number, height: number): Promise<void>;
    fillTriangle(x1: number, y1: number, x2: number, y2: number, x3: number, y3: number): Promise<void>;
    getPixel(x: number, y: number): Promise<string>;
    setPixel(x: number, y: number, color: string): Promise<void>;
    showMessage(text: string, title: string): Promise<void>;
}